various: deploy ACME certs through vault for some frontend jobs
This doesn't really work yet for dependencies like postgres because
I don't yet have a way of deploying the root CA to the cluster for
validation. frontend jobs that have no dependencies like dashboards
are already fair game, though, as it's easy to import the CA cert into
a browser or phone.